+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: flutter_typeahead
-version: 3.1.3
+version: 3.2.0
 
 description: A highly customizable typeahead (autocomplete) text input field for Flutter
 homepage: https://github.com/AbdulRahmanAlHamali/flutter_typeahead
